+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Add check for zero withdraw tasks in manage key value op
 
+### Fixed
+
+* Removing system key values
+
 ## Unreleased
 
 ### Changed