+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 ## Unreleased
 
+### Added
+
+* Return error code if value of withdraw tasks is zero in manage key value op
+
 ### Changed
 
 * Key value resource comparing in rule verifying
 
 ### Fixed
 
+* Removing key values by key which is determined in mValueTypes map in manageKyValueOpFrame
 * Check amount precision fitting on set fee operation level, remove checks for upper and lower bounds
 
 ## 3.0.0-x.5